Add simple intrinsic function with variable argument support

+++ tests/test.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 
-int factorial(int val) {
-  int temp = 1;
+float factorial(int val) {
+  float temp = 1;
   for (int i = 2; i <= val; ++i)
     temp *= i;
   return temp;
